Extract standup_lib.py from weekly script
Move shared GitHub API helpers and activity-gathering functions
into a new .github/scripts/standup_lib.py so the upcoming
on-comment trigger can reuse them without duplicating logic.

Behavior of the weekly Monday post is unchanged: format_contributor
_comment defaults to include_last_week=True, preserving the user
header, @-mention, and "Last Week" link block. The new
include_last_week=False path is reserved for the on-demand
/check-in reply added in a follow-up commit.
